Day 1: Refactor C solution
Looking at the resulting disassembly, this results in a `main` method
that is 16 bytes shorter on x86-64, since the previous conditional
instructions are replaced with a `SUB NEG LEA` instruction sequence.

According to `perf stat`, this also has the effect of reducing the
branch mispredictions (both absolute, and as a percentage of total
branches) encountered, so let's call it a win all round.